+++ workflows_definition/src/openapi/models/changereasonstatusop.py
@@ -5,8 +5,8 @@ from .changereasonstatusreq import ChangeReasonStatusReq, ChangeReasonStatusReqT
 from openapi.types import BaseModel
 from openapi.utils import FieldMetadata, PathParamMetadata, RequestMetadata
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class ChangeReasonStatusRequestTypedDict(TypedDict):