+++ workflows_definition/src/openapi/models/closingreasonid.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 from openapi.types import BaseModel
-from typing import TypedDict
+from typing_extensions import TypedDict
 
 
 class ClosingReasonIDTypedDict(TypedDict):