+++ workflows_definition/src/openapi/models/closingreasons.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 from .closingreason import ClosingReason, ClosingReasonTypedDict
 from openapi.types import BaseModel
-from typing import List, TypedDict
+from typing import List
+from typing_extensions import TypedDict
 
 
 class ClosingReasonsTypedDict(TypedDict):