+++ workflows_definition/src/openapi/models/dynamicduedate.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 from enum import Enum
 from openapi.types import BaseModel
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class ActionTypeCondition(str, Enum):