+++ workflows_definition/src/openapi/models/getworkflowclosingreasonsop.py
@@ -4,8 +4,7 @@ from __future__ import annotations
 from openapi.types import BaseModel
 from openapi.utils import FieldMetadata, PathParamMetadata
 import pydantic
-from typing import TypedDict
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypedDict
 
 
 class GetWorkflowClosingReasonsRequestTypedDict(TypedDict):