+++ workflows_definition/src/openapi/models/section.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 from .itemtype import ItemType
 from .step import Step, StepTypedDict
 from openapi.types import BaseModel
-from typing import List, Optional, TypedDict
-from typing_extensions import NotRequired
+from typing import List, Optional
+from typing_extensions import NotRequired, TypedDict
 
 
 class SectionTypedDict(TypedDict):