+++ workflows_definition/src/openapi/models/security.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 from openapi.types import BaseModel
 from openapi.utils import FieldMetadata, SecurityMetadata
-from typing import TypedDict
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypedDict
 
 
 class SecurityTypedDict(TypedDict):