+++ workflows_definition/src/openapi/models/setworkflowclosingreasonsop.py
@@ -5,8 +5,7 @@ from .closingreasonsids import ClosingReasonsIds, ClosingReasonsIdsTypedDict
 from openapi.types import BaseModel
 from openapi.utils import FieldMetadata, PathParamMetadata, RequestMetadata
 import pydantic
-from typing import TypedDict
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypedDict
 
 
 class SetWorkflowClosingReasonsRequestTypedDict(TypedDict):