+++ workflows_definition/src/openapi/models/step.py
@@ -10,8 +10,8 @@ from .steprequirement import StepRequirement, StepRequirementTypedDict
 from .steptype import StepType
 from openapi.types import BaseModel
 import pydantic
-from typing import List, Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import List, Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class AutomationConfigTypedDict(TypedDict):