+++ workflows_definition/src/openapi/models/updateentityattributes.py
@@ -4,8 +4,7 @@ from __future__ import annotations
 from enum import Enum
 from openapi.types import BaseModel
 import pydantic
-from typing import TypedDict
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypedDict
 
 
 class Source(str, Enum):