+++ workflows_definition/src/openapi/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.7.2"
-    gen_version: str = "2.428.1"
-    user_agent: str = "speakeasy-sdk/python 1.7.2 2.428.1 1.0.0 openapi"
+    sdk_version: str = "1.8.0"
+    gen_version: str = "2.429.0"
+    user_agent: str = "speakeasy-sdk/python 1.8.0 2.429.0 1.0.0 openapi"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None